be freq_t type agnostic
git-svn-id: https://hamlib.svn.sourceforge.net/svnroot/hamlib/trunk@1651 7ae35d74-ebe9-4afe-98af-79ac388436b8

+++ tests/testfreq.c
@@ -32,27 +32,27 @@ int main (int argc, char *argv[])
 
 	/* freq on 31bits test */
 	f = GHz(2);
-	printf("GHz(2) = %lld\n", f);
+	printf("GHz(2) = %lld\n", (long long)f);
 
 	/* freq on 32bits test */
 	f = GHz(4);
-	printf("GHz(4) = %lld\n", f);
+	printf("GHz(4) = %lld\n", (long long)f);
 
 	/* freq on >32bits test */
 	f = GHz(5);
-	printf("GHz(5) = %lld\n", f);
+	printf("GHz(5) = %lld\n", (long long)f);
 
 	/* floating point to freq conversion test */
 	f = GHz(1.3);
-	printf("GHz(1.3) = %lld\n", f);
+	printf("GHz(1.3) = %lld\n", (long long)f);
 
 	/* floating point to freq conversion precision test */
 	f = GHz(1.234567890);
-	printf("GHz(1.234567890) = %lld\n", f);
+	printf("GHz(1.234567890) = %lld\n", (long long)f);
 
 	/* floating point to freq conversion precision test, with freq >32bits */
 	f = GHz(123.456789012);
-	printf("GHz(123.456789012) = %lld\n", f);
+	printf("GHz(123.456789012) = %lld\n", (long long)f);
 
 	return 0;
 }
